CXP-1295: rename
