+++ components/catalogs/back/src/Application/Handler/GetProductMappingSchemaHandler.php
@@ -8,13 +8,13 @@ use Akeneo\Catalogs\Application\Exception\CatalogNotFoundException;
 use Akeneo\Catalogs\Application\Persistence\Catalog\GetCatalogQueryInterface;
 use Akeneo\Catalogs\Application\Storage\CatalogsMappingStorageInterface;
 use Akeneo\Catalogs\ServiceAPI\Exception\CatalogNotFoundException as ServiceApiCatalogNotFoundException;
-use Akeneo\Catalogs\ServiceAPI\Query\GetCatalogProductMappingSchemaQuery;
+use Akeneo\Catalogs\ServiceAPI\Query\GetProductMappingSchemaQuery;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class GetCatalogProductMappingSchemaHandler
+final class GetProductMappingSchemaHandler
 {
     public function __construct(
         private GetCatalogQueryInterface $getCatalogQuery,
@@ -26,7 +26,7 @@ final class GetCatalogProductMappingSchemaHandler
      * @psalm-suppress MixedReturnStatement
      * @psalm-suppress MixedInferredReturnType
      */
-    public function __invoke(GetCatalogProductMappingSchemaQuery $query): object
+    public function __invoke(GetProductMappingSchemaQuery $query): object
     {
         try {
             $catalog = $this->getCatalogQuery->execute($query->getCatalogId());
@@ -34,14 +34,14 @@ final class GetCatalogProductMappingSchemaHandler
             throw new ServiceApiCatalogNotFoundException();
         }
 
-        $catalogProductMappingSchema = \stream_get_contents(
+        $productMappingSchema = \stream_get_contents(
             $this->catalogsMappingStorage->read(\sprintf('%d_product.json', $catalog->getId()))
         );
 
-        if (false === $catalogProductMappingSchema) {
+        if (false === $productMappingSchema) {
             throw new \LogicException('Product mapping schema is unreadable.');
         }
 
-        return \json_decode($catalogProductMappingSchema, false, 512, JSON_THROW_ON_ERROR);
+        return \json_decode($productMappingSchema, false, 512, JSON_THROW_ON_ERROR);
     }
 }
