+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/routing/public.yml
@@ -40,5 +40,5 @@ akeneo_catalogs_public_get_products:
 
 akeneo_catalogs_public_update_products_mapping_schema:
     path: '/api/rest/v1/catalogs/{catalogId}/mapping-schemas/product'
-    controller: Akeneo\Catalogs\Infrastructure\Controller\Public\UpdateCatalogProductMappingSchemaAction
+    controller: Akeneo\Catalogs\Infrastructure\Controller\Public\UpdateProductMappingSchemaAction
     methods: [ PUT ]
